feature/refactor-easy-random (#1227)
