+++ packages/EasyRandom/src/Bridge/Symfony/Resources/config/definition.php
@@ -1,0 +1,18 @@
+<?php
+declare(strict_types=1);
+
+use Symfony\Component\Config\Definition\Configurator\DefinitionConfigurator;
+
+return static function (DefinitionConfigurator $definition) {
+    $definition->rootNode()
+        ->children()
+            ->integerNode('uuid_version')
+                ->validate()
+                    ->ifNotInArray([4, 6])
+                    ->thenInvalid('Invalid UUID version %s')
+                ->end()
+                ->defaultValue(6)
+                ->info('Version of UUID to generate')
+            ->end()
+        ->end();
+};
